requirements.txt
Django==1.7.1
IPy==0.81
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.18
celery==3.1.16
django-widget-tweaks==1.3
djorm-pgarray==1.1
dnspython3==1.12.0
geoip2==2.0.2
html5lib==0.999
ipwhois==0.9.1
jsonfield==1.0.0
kombu==3.0.23
requests==2.3.0
six==1.7.3
psycopg2==2.5.4
pythonwhois==2.4.3
python-dateutil==2.4.0
pytz==2014.9
robobrowser==0.5.1
SQLAlchemy==0.9.8
tldextract==1.5.1
werkzeug==0.9.6